requirements.txt
supabase 
pandas 
beautifulsoup4 
playwright
psycopg2-binary